rename contexts file
